requirements.txt
## requirements.txt
psycopg2-binary==2.9.7
python-dotenv==1.0.0
PySide6>=6.5.0
bcrypt>=4.0.0
# sqlite3